Test subscribe branches without wildcards, multiple projects
