+++ src/test/java/com/ribose/jenkins/plugin/awscodecommittrigger/it/fixture/ProjectFixture.java
@@ -23,22 +23,21 @@ import com.google.gson.GsonBuilder;
 public class ProjectFixture {
     private static final Long TIMEOUT = 60_000L;//in milliseconds, e.g: 300_000 ~ 5 mins
 
-    private String listenBranches;
+    private String subscribedBranches;
     private String[] sendBranches;
     private Boolean shouldStarted;
     private Long timeout = TIMEOUT;
-    private String gitUrl = "https://git-codecommit.us-west-2.amazonaws.com/v1/repos/testjenkins";
 
     private static final Gson gson = new GsonBuilder()
         .setFieldNamingStrategy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
         .create();
 
-    public String getListenBranches() {
-        return listenBranches;
+    public String getSubscribedBranches() {
+        return subscribedBranches;
     }
 
-    public ProjectFixture setListenBranches(String listenBranches) {
-        this.listenBranches = listenBranches;
+    public ProjectFixture setSubscribedBranches(String subscribedBranches) {
+        this.subscribedBranches = subscribedBranches;
         return this;
     }
 
@@ -66,15 +65,6 @@ public class ProjectFixture {
 
     public ProjectFixture setTimeout(Long timeout) {
         this.timeout = timeout;
-        return this;
-    }
-
-    public String getGitUrl() {
-        return gitUrl;
-    }
-
-    public ProjectFixture setGitUrl(String gitUrl) {
-        this.gitUrl = gitUrl;
         return this;
     }
